Fix a bug with slideritem.
When there was only one item in the slider, template { for } crashed
with 'object is not iterable'.
The template logic worked only with iterables, so fixed that.
Such pages as news or blog entries didn't work while petitions
and the main page did.
There were earlier problems with the slider (on the main page),
which disappeared by themselves, not sure whether this will improve that.